Added pretty age picker
Added Exportable functionality to selector lists using the provider class.
Now when an age is updated, the API will deconstruct it and add both the min and max age to the exportable provider that is created at the start of the application.
Same for the budget.
Starting on gender selection next.